requirements.txt
# ---------------------------------------------------------------------------
# ProofPixel — Backend Dependencies
# ---------------------------------------------------------------------------
# Install with:  pip install -r requirements.txt
# ---------------------------------------------------------------------------

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ML / Computer Vision
opencv-python-headless>=4.9.0
scikit-image>=0.22.0
scikit-learn>=1.4.0
joblib>=1.3.0
numpy>=1.26.0

# Database
mysql-connector-python>=8.3.0

# Environment variables
python-dotenv>=1.0.0

# Upload handling (FastAPI dependency)
python-multipart>=0.0.9

# Training utilities
tqdm>=4.66.0
